requirements.txt
streamlit
groq
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
extra-streamlit-components
